Scraping complete
Seems to be working well now. Had to squash some bugs and CSV writing was a pain in the ass for some reason but all seems good. Glad I erred on the side of caution running the script with small date ranges to begin as it's pretty intensive, don't want to crash their servers. Will probably want to add the ability to retrieve only specific days (eg Saturday) to minimize impact for them. Won't be hard as I'm already getting day values. Now what to do with all this data...